Bugfixes for computers with multiple Steam libraries, and a problem with some game names.
I keep forgetting to actually sync the module as well as just committing
it. :shit:
This fixes a bug where the module would display an error when run
against a computer with multiple Steam libraries. It also fixes a bug
caused by that fix, where the module would not get some
games/applications (e.g. You Need A Budget) and would only get one entry
if a common word was entered (e.g. "saints" was only getting Saints Row:
The Third, not IV or Gat out of Hell).